feat: Solve 2023/01
We are off again! So annoying though how I would find the first instance of a digit, even if there was one later of the same in the same line. Envious of those compact solutions, but they would be heck to troubleshoot, would they not?

I hated writing the my_find_all function, but it had become late and I really did not want to introduce regex. Should read the megathread for hints on how to do all this neatly